+++ Private/Device/ExportPanDeviceDb.ps1
@@ -43,22 +43,23 @@ PowerPAN private helper function to serialize and store PanDevice objects from P
       # Aggregate devices through each iteration of Process block (and foreach within Process block itself)
       # .NET Generic List provides under-the-hood efficiency during add/remove compared to PowerShell native arrays or ArrayList.
       # List type is [PSObject] -- objects are changed from [PanDevice] to custom objects on the way out to export.
-      $StoredDeviceAgg = [System.Collections.Generic.List[PSObject]]@()
+      $StoredDeviceAgg = [System.Collections.Generic.List[ordered]]@()
    } # Begin Block
 
    Process {
-      foreach ($DeviceCur in $Global:PanDeviceDb) {
+      foreach ($DeviceCur in ($Global:PanDeviceDb | Where-Object {$_.Persist -eq $true} )) {
          Write-Debug ('{0}: Name: {1}' -f $MyInvocation.MyCommand.Name,$DeviceCur.Name)
          # Build a custom hash table suitable for serializing to JSON in PowerShell End block
          # *Cannot* serialize the raw [PanDevice] object direclty. Needs some massaging to get
          # the Credential.Password [SecureString] and Key [SecureString] to their encrypted serializable/storable form.
-         $CustomObj = @{
-            'Name' = $DeviceCur.Name;
-            'Username' = $DeviceCur.Credential.UserName;
-            'ValidateCertificate' = $DeviceCur.ValidateCertificate;
-            'Protocol' = $DeviceCur.Protocol;
-            'Port' = $DeviceCur.Port;
-            'Type' = $DeviceCur.Type
+         $CustomObj = [ordered]@{
+            'Name' =                $DeviceCur.Name
+            'Username' =            $DeviceCur.Credential.UserName
+            'ValidateCertificate' = $DeviceCur.ValidateCertificate
+            'Protocol' =            $DeviceCur.Protocol
+            'Port' =                $DeviceCur.Port
+            'Type' =                $DeviceCur.Type
+            'Location' =            $DeviceCur.Location
          } # End hash table
 
          # Add the Credential.Password
@@ -97,7 +98,7 @@ PowerPAN private helper function to serialize and store PanDevice objects from P
       if(-not [String]::IsNullorEmpty($StoredDeviceAgg) ) {
          Write-Debug ('{0}: Storing {1} device{2} to {3}' -f $MyInvocation.MyCommand.Name,$StoredDeviceAgg.Count,$(if($StoredDeviceAgg.Count -ne 1){'s'}),$StoredJsonPath)
          # Serialize and write to storage
-         ConvertTo-Json -InputObject $StoredDeviceAgg | Set-Content -Path $StoredJsonPath -Force | Out-Null
+         ConvertTo-Json -InputObject $StoredDeviceAgg -Depth 10 | Set-Content -Path $StoredJsonPath -Force | Out-Null
       }
    } # End Block
 } # Function