+++ Public/Device/Get-PanDevice.ps1
@@ -10,28 +10,6 @@ Multiple -Name (via array) is a logical OR match with regular expression support
 
 When -Session, -Label, -Name are used together, specific parameter AND/OR match behavior
 still applies, but each parameter in use must match (AND across all in use parameters)
-
-:: Location Property Update ::
-Each PanDevice has a Location property holding shared, vsys, and device group "Name to XPath"
-mappings for the device. Location property is *not* persisted to disk or preserved across
-PowerShell sessions and is updated dynamically during the PowerShell session.
-
-Get-PanDevice is *also* the principal trigger for updating the PanDevice.Location property.
-
-Performing the update requires XML API calls to the device retrieve its current shared, vsys1,
-vsys2, device group, etc. layout.
-
-Why does this matter? Imagine:
-   - 10 PanDevice's persisting across PowerShell sessions
-   - 8 them are network reachable by Name/IP from your current workstation
-   - 2 of them are NOT network reachable by Name/IP from your current workstation
-   - A call to "Get-PanDevice -All" in a *new* PowerShell session will attempt to refresh all 10
-   - 8 device's Location properties will update very quickly
-   - 2 unreachable devices will hang for many seconds while attempting to reach their API's
-
-Recommendation:
-   - Use "Get-PanDevice -All -NoLocation" to "see" all your PanDevice's without updating Location
-   - When wanting to get a device for actual use, scope Get-PanDevice with -Name or -Label
 .INPUTS
 None
 .OUTPUTS
@@ -83,9 +61,7 @@ If $Global:PanDeviceLabelDefault is empty, returns PanDevice(s) created in curre
       [parameter(Mandatory=$false,ParameterSetName='Filter',HelpMessage='Case-insensitive exact match for Label. Multiple Label is logical AND match')]
       [String[]] $Label,
       [parameter(Mandatory=$true,ParameterSetName='All',HelpMessage='Switch parameter for ALL PanDevice')]
-      [Switch] $All,
-      [parameter(Mandatory=$false,HelpMessage='Do not update PanDevice Location map')]
-      [Switch] $NoLocation
+      [Switch] $All
    )
 
    # Propagate -Debug and -Verbose to this module function, https://tinyurl.com/y5dcbb34
@@ -205,13 +181,6 @@ If $Global:PanDeviceLabelDefault is empty, returns PanDevice(s) created in curre
       Write-Debug ($MyInvocation.MyCommand.Name + ': All ParameterSetName')
       $DeviceAgg = $Global:PanDeviceDb
    } # ParameterSetName 'All'
-
-   # Ensure Location map is up to date for PanDevice's being returned
-   if(-not $PSBoundParameters.NoLocation.IsPresent) {
-      foreach($DeviceCur in $DeviceAgg) {
-         Update-PanDeviceLocation -Device $DeviceCur
-      }
-   }
 
    return $DeviceAgg
 } # Function