+++ Public/Object/Move-PanObject.ps1
@@ -111,6 +111,9 @@ PanServiceGroup
             Write-Debug ('{0} (as {1}): Device: {2} Location: {3} Name: {4} DstLocation: {5} ' -f 
                $MyInvocation.MyCommand.Name, $MyInvocation.InvocationName, $DeviceCur.Name, $PSBoundParameters.Location,
                $PSBoundParameters.Name, $PSBoundParameters.DstLocation)
+            # Update Location if past due
+            if($PSBoundParameters.Device.LocationUpdated.AddSeconds($Global:PanDeviceLocRefSec) -lt (Get-Date)) { Update-PanDeviceLocation -Device $PSBoundParameters.Device }
+            
             # Given -Device ParameterSet, fetch the object for its XPath
             switch ($MyInvocation.InvocationName) {
                'Move-PanAddress'       { $Obj = Get-PanAddress -Device $DeviceCur -Location $PSBoundParameters.Location -Name $PSBoundParameters.Name; continue }